attack_framework/OpenRT-main/requirements.txt
# Core Dependencies
openai>=2.8.1
openai-agents>=0.6.1
nest-asyncio>=1.6.0
aiofiles>=25.1.0
python-dotenv>=1.2.1

# Data Processing
pandas>=2.2.3

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# Tokenization
protobuf>=6.0.0
tiktoken>=0.12.0
sentencepiece>=0.2.0

# Image Processing
Pillow>=12.0.0

# Progress & Utilities
tqdm>=4.67.1

# Machine Learning & AI
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0

# NLP & Embeddings
sentence-transformers>=2.7.0
spacy>=3.7.0
nltk>=3.8.0

# ASCII Art & Text Processing
art>=6.0.0